requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.6.1
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.15.0
jupyter>=1.0.0
mlflow>=2.8.0
streamlit>=1.28.0

# MongoDB integration
pymongo>=4.6.0
python-dotenv>=1.0.0

# Deep Learning libraries
tensorflow>=2.13.0
keras>=2.13.0

# Additional clustering and visualization libraries
scipy>=1.11.0
yellowbrick>=1.5
prince>=0.7.0  # For PCA and other dimensionality reduction
umap-learn>=0.5.0  # For advanced dimensionality reduction

# Model persistence and deployment
joblib>=1.3.0
dill>=0.3.7

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0